+++ modules/actions/snapshots/lib/apply-usd-values.ts
@@ -16,53 +16,68 @@ export const applyUSDValues = async (
 
     const poolIds = [...new Set(rawSnapshots.map((snapshot) => snapshot.poolId))];
 
-    // Pool tokens are needed, because SG returns raw token amounts
+    // Pool tokens are needed, because SG returns raw token amounts and we need token addresses
     const poolTokens = await fetchPoolTokens(poolIds);
 
-    // For each timestamp, fetch the prices and calculate USD values
-    const groupedTimestamps = _.groupBy(rawSnapshots, 'timestamp');
+    const groupedByPoolId = _.groupBy(rawSnapshots, 'poolId');
 
-    for (const [timestamp, group] of Object.entries(groupedTimestamps)) {
-        // For each poolId, calculate USD values
-        const prices = await fetchPrices(
-            group[0].chain,
-            parseInt(timestamp) < lastMidnight ? parseInt(timestamp) : undefined,
-        );
+    let snapshotCount = 0;
+    for (const [, poolSnapshots] of Object.entries(groupedByPoolId)) {
+        console.log('Processing snapshots for pool', poolSnapshots[0].poolId, poolSnapshots[0].chain);
+        const tokens = _.sortBy(poolTokens[poolSnapshots[0].poolId], 'index');
 
-        for (const snapshot of group) {
-            if (
-                !poolTokens[snapshot.poolId] ||
-                poolTokens[snapshot.poolId].length === 0 ||
-                !snapshot.amounts ||
-                !snapshot.totalVolumes ||
-                !snapshot.totalSwapFees
-            ) {
-                snapshots.push(snapshot);
-                continue;
+        if (!tokens) {
+            throw new Error(
+                `Pool tokens not found for pool ${poolSnapshots[0].poolId} on chain ${poolSnapshots[0].chain}`,
+            );
+        }
+
+        const sortedSnapshots = _.sortBy(poolSnapshots, 'timestamp');
+        for (let index = 0; index < sortedSnapshots.length; index++) {
+            const snapshot = sortedSnapshots[index];
+            const previousSnapshot = index > 0 && snapshots[snapshotCount - 1];
+
+            const prices = await fetchPrices(
+                snapshot.chain,
+                snapshot.timestamp < lastMidnight ? snapshot.timestamp : undefined,
+            );
+
+            let volume24h = 0;
+            const swapTokenWithPrice = Object.values(tokens).find(({ address }) => prices[address]);
+            if (swapTokenWithPrice) {
+                const swapVolume = (snapshot.dailyVolumes as string[])[swapTokenWithPrice.index] || '0'; // Some snapshots have empty arrays
+                volume24h = parseFloat(swapVolume) * prices[swapTokenWithPrice.address];
             }
 
-            const tokens = _.keyBy(poolTokens[snapshot.poolId], 'index');
-
-            // With token do we have the price for?
-            let swapTokenIndex = Object.values(tokens).findIndex(({ address }) => prices[address]);
-            if (swapTokenIndex < 0) swapTokenIndex = 0;
-            const swapVolume = (snapshot.totalVolumes as string[])[swapTokenIndex];
-
-            // Swap volume is only for the tokenIn
-            const totalSwapVolume = parseFloat(swapVolume) * (prices[tokens[swapTokenIndex].address] || 0);
+            const fees24h = calculateValue(snapshot.dailySwapFees as string[], tokens, prices);
+            const surplus24h = calculateValue(snapshot.dailySurpluses as string[], tokens, prices);
             const totalLiquidity = calculateValue(snapshot.amounts as string[], tokens, prices);
-            const totalSwapFee = calculateValue(snapshot.totalSwapFees as string[], tokens, prices);
             const sharePrice = snapshot.totalSharesNum === 0 ? 0 : totalLiquidity / snapshot.totalSharesNum;
 
-            // Calculate USD values
-            const usdValues = {
+            let totalSwapVolume = snapshot.totalSwapVolume || volume24h;
+            let totalSwapFee = snapshot.totalSwapFee || fees24h;
+            let totalSurplus = snapshot.totalSurplus || surplus24h;
+
+            // Calculate daily values as the difference from the previous snapshot
+            if (previousSnapshot) {
+                totalSwapVolume = previousSnapshot.totalSwapVolume + volume24h;
+                totalSwapFee = previousSnapshot.totalSwapFee + fees24h;
+                totalSurplus = (previousSnapshot.totalSurplus || 0) + surplus24h;
+            }
+
+            snapshots.push({
+                ...snapshot,
+                volume24h,
+                fees24h,
+                surplus24h,
                 totalLiquidity,
+                sharePrice,
                 totalSwapVolume,
                 totalSwapFee,
-                sharePrice,
-            };
+                totalSurplus,
+            });
 
-            snapshots.push({ ...snapshot, ...usdValues });
+            snapshotCount++;
         }
     }
 
@@ -70,6 +85,8 @@ export const applyUSDValues = async (
 };
 
 const calculateValue = (amounts: string[], tokens: Record<number, any>, prices: Record<string, number>) => {
+    if (!amounts || !tokens || !prices) return 0;
+
     return amounts.reduce((acc, amount, index) => {
         const token = tokens[index];
         return token && prices[token.address] ? acc + parseFloat(amount) * prices[token.address] : acc;
@@ -86,9 +103,7 @@ const fetchPricesHelper = async (chain: Chain, timestamp?: number): Promise<Reco
     // Check cache
     const cacheKey = `prices-${chain}-${timestamp || 'current'}`;
     const cachedPrices = cache.get(cacheKey);
-    if (cachedPrices) {
-        return cachedPrices;
-    }
+    if (cachedPrices) return cachedPrices;
 
     const selector = {
         where: { chain, ...(timestamp ? { timestamp } : {}) }, // No timestamp for current prices