+++ modules/actions/snapshots/lib/compute-daily-values.spec.ts
@@ -1,188 +1,130 @@
 import { computeDailyValues } from './compute-daily-values';
 import { Prisma } from '@prisma/client';
 
+const defaultSnapshot = {
+    poolId: '1',
+    timestamp: 1 * 86400,
+    totalVolumes: ['100', '200'],
+    totalSwapFees: ['10', '20'],
+    totalSurpluses: ['5', '15'],
+};
+
 describe('computeDailyValues', () => {
-    it('should compute inital values for a single snapshot', () => {
-        const inputSnapshots = [{ id: '1', poolId: 'poolA', timestamp: 86400, totalSwapVolume: 100, totalSwapFee: 10 }];
-
-        const expectedOutput = [
-            {
-                id: '1',
-                poolId: 'poolA',
-                timestamp: 86400,
-                totalSwapVolume: 100,
-                totalSwapFee: 10,
-                volume24h: 100,
-                fees24h: 10,
-            },
-        ];
-
-        const result = computeDailyValues(inputSnapshots as Prisma.PrismaPoolSnapshotUncheckedCreateInput[]);
-
-        expect(result).toEqual(expectedOutput);
-    });
-    it('should compute daily values for a single pool', () => {
-        const inputSnapshots = [
-            { id: '1', poolId: 'poolA', timestamp: 86400, totalSwapVolume: 100, totalSwapFee: 10 },
-            {
-                id: '2',
-                poolId: 'poolA',
-                timestamp: 2 * 86400,
-                totalSwapVolume: 300,
-                totalSwapFee: 30,
-            },
-            {
-                id: '3',
-                poolId: 'poolA',
-                timestamp: 3 * 86400,
-                totalSwapVolume: 600,
-                totalSwapFee: 50,
-            },
-        ];
-
-        const expectedOutput = [
-            {
-                id: '1',
-                poolId: 'poolA',
-                timestamp: 86400,
-                totalSwapVolume: 100,
-                totalSwapFee: 10,
-                volume24h: 0,
-                fees24h: 0,
-            },
-            {
-                id: '2',
-                poolId: 'poolA',
-                timestamp: 2 * 86400,
-                totalSwapVolume: 300,
-                totalSwapFee: 30,
-                volume24h: 200,
-                fees24h: 20,
-            },
-            {
-                id: '3',
-                poolId: 'poolA',
-                timestamp: 3 * 86400,
-                totalSwapVolume: 600,
-                totalSwapFee: 50,
-                volume24h: 300,
-                fees24h: 20,
-            },
-        ];
-
-        const result = computeDailyValues(inputSnapshots as Prisma.PrismaPoolSnapshotUncheckedCreateInput[]);
-
-        expect(result).toEqual(expectedOutput);
-    });
-
-    it('should handle multiple pools independently', () => {
-        const inputSnapshots = [
-            { id: '1', poolId: 'poolA', timestamp: 86400, totalSwapVolume: 100, totalSwapFee: 10 },
-            {
-                id: '2',
-                poolId: 'poolA',
-                timestamp: 2 * 86400,
-                totalSwapVolume: 300,
-                totalSwapFee: 30,
-            },
-            { id: '3', poolId: 'poolB', timestamp: 86400, totalSwapVolume: 200, totalSwapFee: 20 },
-            {
-                id: '4',
-                poolId: 'poolB',
-                timestamp: 2 * 86400,
-                totalSwapVolume: 500,
-                totalSwapFee: 50,
-            },
-        ];
-
-        const expectedOutput = [
-            {
-                id: '1',
-                poolId: 'poolA',
-                timestamp: 86400,
-                totalSwapVolume: 100,
-                totalSwapFee: 10,
-                volume24h: 0,
-                fees24h: 0,
-            },
-            {
-                id: '2',
-                poolId: 'poolA',
-                timestamp: 2 * 86400,
-                totalSwapVolume: 300,
-                totalSwapFee: 30,
-                volume24h: 200,
-                fees24h: 20,
-            },
-            {
-                id: '3',
-                poolId: 'poolB',
-                timestamp: 86400,
-                totalSwapVolume: 200,
-                totalSwapFee: 20,
-                volume24h: 0,
-                fees24h: 0,
-            },
-            {
-                id: '4',
-                poolId: 'poolB',
-                timestamp: 2 * 86400,
-                totalSwapVolume: 500,
-                totalSwapFee: 50,
-                volume24h: 300,
-                fees24h: 30,
-            },
-        ];
-
-        const result = computeDailyValues(inputSnapshots as Prisma.PrismaPoolSnapshotUncheckedCreateInput[]);
-
-        expect(result).toEqual(expectedOutput);
-    });
-    it('should handle missing totalSwapVolume and totalSwapFee gracefully', () => {
-        const inputSnapshots = [
-            {
-                id: '2',
-                poolId: 'poolA',
-                timestamp: 2 * 86400,
-                totalSwapVolume: 200,
-                totalSwapFee: 20,
-                volume24h: 100,
-                fees24h: 10,
-            },
-            { id: '3', poolId: 'poolA', timestamp: 3 * 86400, totalSwapVolume: 400, totalSwapFee: 40 },
-        ];
-
-        const expectedOutput = [
-            {
-                id: '2',
-                poolId: 'poolA',
-                timestamp: 2 * 86400,
-                totalSwapVolume: 200,
-                totalSwapFee: 20,
-                volume24h: 100,
-                fees24h: 10,
-            },
-            {
-                id: '3',
-                poolId: 'poolA',
-                timestamp: 3 * 86400,
-                totalSwapVolume: 400,
-                totalSwapFee: 40,
-                volume24h: 200,
-                fees24h: 20,
-            },
-        ];
-
-        const result = computeDailyValues(inputSnapshots as Prisma.PrismaPoolSnapshotUncheckedCreateInput[]);
-
-        expect(result).toEqual(expectedOutput);
-    });
-
     it('should return an empty array when input is empty', () => {
-        const inputSnapshots = [];
-
-        const result = computeDailyValues(inputSnapshots);
-
+        const result = computeDailyValues([]);
         expect(result).toEqual([]);
+    });
+
+    it('should handle a single snapshot per pool correctly', () => {
+        const snapshots = [defaultSnapshot];
+        const result = computeDailyValues(snapshots as Prisma.PrismaPoolSnapshotUncheckedCreateInput[]);
+        expect(result).toEqual([
+            {
+                ...defaultSnapshot,
+                dailyVolumes: ['100', '200'],
+                dailySwapFees: ['10', '20'],
+                dailySurpluses: ['5', '15'],
+            },
+        ]);
+    });
+
+    it('should calculate daily values for multiple snapshots of a single pool', () => {
+        const snapshots = [
+            { ...defaultSnapshot },
+            {
+                ...defaultSnapshot,
+                timestamp: 2 * 86400,
+                totalVolumes: ['150', '250'],
+                totalSwapFees: ['15', '25'],
+                totalSurpluses: ['10', '20'],
+            },
+        ];
+        const result = computeDailyValues(snapshots as Prisma.PrismaPoolSnapshotUncheckedCreateInput[]);
+        expect(result).toEqual([
+            {
+                ...defaultSnapshot,
+                dailyVolumes: ['100', '200'],
+                dailySwapFees: ['10', '20'],
+                dailySurpluses: ['5', '15'],
+            },
+            {
+                ...defaultSnapshot,
+                timestamp: 2 * 86400,
+                totalVolumes: ['150', '250'],
+                totalSwapFees: ['15', '25'],
+                totalSurpluses: ['10', '20'],
+                dailyVolumes: ['50', '50'],
+                dailySwapFees: ['5', '5'],
+                dailySurpluses: ['5', '5'],
+            },
+        ]);
+    });
+
+    it('should calculate daily values for multiple pools', () => {
+        const snapshots = [
+            { ...defaultSnapshot },
+            {
+                ...defaultSnapshot,
+                poolId: '2',
+                totalVolumes: ['200', '400'],
+                totalSwapFees: ['20', '40'],
+                totalSurpluses: ['10', '30'],
+            },
+            {
+                ...defaultSnapshot,
+                timestamp: 2 * 86400,
+                totalVolumes: ['150', '250'],
+                totalSwapFees: ['15', '25'],
+                totalSurpluses: ['10', '20'],
+            },
+            {
+                ...defaultSnapshot,
+                poolId: '2',
+                timestamp: 2 * 86400,
+                totalVolumes: ['250', '450'],
+                totalSwapFees: ['25', '45'],
+                totalSurpluses: ['15', '35'],
+            },
+        ];
+        const result = computeDailyValues(snapshots as Prisma.PrismaPoolSnapshotUncheckedCreateInput[]);
+        expect(result).toEqual([
+            {
+                ...defaultSnapshot,
+                dailyVolumes: ['100', '200'],
+                dailySwapFees: ['10', '20'],
+                dailySurpluses: ['5', '15'],
+            },
+            {
+                ...defaultSnapshot,
+                timestamp: 2 * 86400,
+                totalVolumes: ['150', '250'],
+                totalSwapFees: ['15', '25'],
+                totalSurpluses: ['10', '20'],
+                dailyVolumes: ['50', '50'],
+                dailySwapFees: ['5', '5'],
+                dailySurpluses: ['5', '5'],
+            },
+            {
+                ...defaultSnapshot,
+                poolId: '2',
+                totalVolumes: ['200', '400'],
+                totalSwapFees: ['20', '40'],
+                totalSurpluses: ['10', '30'],
+                dailyVolumes: ['200', '400'],
+                dailySwapFees: ['20', '40'],
+                dailySurpluses: ['10', '30'],
+            },
+            {
+                ...defaultSnapshot,
+                poolId: '2',
+                timestamp: 2 * 86400,
+                totalVolumes: ['250', '450'],
+                totalSwapFees: ['25', '45'],
+                totalSurpluses: ['15', '35'],
+                dailyVolumes: ['50', '50'],
+                dailySwapFees: ['5', '5'],
+                dailySurpluses: ['5', '5'],
+            },
+        ]);
     });
 });