+++ modules/actions/snapshots/pool-snapshot-service.ts
@@ -341,6 +341,11 @@ export class PoolSnapshotService {
             totalVolumes: [],
             totalSwapFees: [],
             totalSurpluses: [],
+            dailyProtocolSwapFees: [],
+            dailyProtocolYieldFees: [],
+            dailyVolumes: [],
+            dailySwapFees: [],
+            dailySurpluses: [],
         };
     }