+++ modules/sources/subgraphs/cow-amm/index.ts
@@ -10,7 +10,10 @@ import {
     PoolShareFragment,
     PoolShare_OrderBy,
     PoolSharesQueryVariables,
+    PoolSnapshot_Filter,
 } from './generated/types';
+import { snapshotToDb } from './transformers/snapshotsToDb';
+import { Chain, Prisma } from '@prisma/client';
 
 /**
  * Builds a client based on subgraph URL.
@@ -18,12 +21,19 @@ import {
  * @param subgraphUrl - url of the subgraph
  * @returns sdk - generated sdk for the subgraph
  */
-export const getCowAmmSubgraphClient = (subgraphUrl: string) => {
+export const getCowAmmSubgraphClient = (subgraphUrl: string, chain: Chain) => {
     const client = new GraphQLClient(subgraphUrl);
     const sdk = getSdk(client);
 
     return {
         ...sdk,
+        async getMetadata() {
+            const { meta } = await sdk.Metadata();
+            if (!meta) {
+                throw new Error('Missing meta data');
+            }
+            return meta;
+        },
         async getAllPools(where: PoolsQueryVariables['where']): Promise<CowAmmPoolFragment[]> {
             const limit = 1000;
             let hasMore = true;
@@ -98,6 +108,31 @@ export const getCowAmmSubgraphClient = (subgraphUrl: string) => {
             }
 
             return snapshots;
+        },
+        async getAllSnapshots(where: PoolSnapshot_Filter): Promise<Prisma.PrismaPoolSnapshotUncheckedCreateInput[]> {
+            const limit = 1000;
+            let hasMore = true;
+            let id = `0x`;
+            let snapshots: CowAmmSnapshotFragment[] = [];
+
+            while (hasMore) {
+                const response = await sdk.Snapshots({
+                    where: { ...where, id_gt: id },
+                    orderBy: PoolSnapshot_OrderBy.Id,
+                    orderDirection: OrderDirection.Asc,
+                    first: limit,
+                });
+
+                snapshots = [...snapshots, ...response.poolSnapshots];
+
+                if (response.poolSnapshots.length < limit) {
+                    hasMore = false;
+                } else {
+                    id = snapshots[snapshots.length - 1].id;
+                }
+            }
+
+            return snapshots.map((s) => snapshotToDb(chain, 1, s));
         },
     };
 };